chore: scope lint and test commands
