+++ .github/workflows/ci.yml
@@ -16,5 +16,6 @@ jobs:
         with:
           go-version: '1.24.x'
       - run: go mod tidy
-      - run: go test ./...
-      - run: go vet ./...
+      - run: make lint
+      - run: make test
+      - run: make vet
