+++ internal/clipboard/auto_linux.go
@@ -3,6 +3,11 @@
 
 package clipboard
 
+import (
+	"os"
+	"strings"
+)
+
 func autoBackendCandidates() []Backend {
 	var candidates []Backend
 	if isWayland() {
@@ -15,4 +20,15 @@ func autoBackendCandidates() []Backend {
 		return nil
 	}
 	return candidates
+}
+
+func isWayland() bool {
+	if v := strings.TrimSpace(os.Getenv("XDG_SESSION_TYPE")); strings.EqualFold(v, "wayland") {
+		return true
+	}
+	return strings.TrimSpace(os.Getenv("WAYLAND_DISPLAY")) != ""
+}
+
+func isX11() bool {
+	return strings.TrimSpace(os.Getenv("DISPLAY")) != ""
 }
